refactor(client): extract shared components flagged by fallow dupes
- LoginGoogle/LoginOrcid (identical except provider string) merged into
  LoginOauthCallback with a providerName route param
- SubmissionsFilterPanelRoles/Status reduced to thin wrappers around new
  SubmissionsFilterPanelOptions; option lists and defaultOptions exports
  unchanged
- InlineCommentReply/OverallCommentReply merged into CommentReply with a
  commentType prop; data-cy values preserved for cypress
- users/beta-users admin tables share userColumns.ts base columns

CommentReply uses a PascalCase template tag: kebab <comment-reply>
camelizes to the existing commentReply ref and shadows the import.
