+++ client/src/pages/Admin/components/SubmissionsFilterPanelStatus.vue
@@ -1,48 +1,16 @@
 <template>
-  <q-card-section class="q-pb-none">
-    <div class="row items-center justify-between q-mb-xs">
-      <div class="text-weight-bold">
-        {{ $t("submissions.filters.status_header") }}
-      </div>
-      <q-btn-group flat>
-        <q-btn
-          dense
-          flat
-          no-caps
-          size="sm"
-          :label="$t('admin.filters.all')"
-          @click="selectAll"
-        />
-        <q-btn
-          dense
-          flat
-          no-caps
-          size="sm"
-          :label="$t('admin.filters.none')"
-          @click="selectNone"
-        />
-        <q-btn
-          dense
-          flat
-          no-caps
-          size="sm"
-          :label="$t('admin.filters.invert')"
-          @click="invert"
-        />
-      </q-btn-group>
-    </div>
-    <q-option-group v-model="filter" :options="tOptions" type="checkbox" />
-  </q-card-section>
+  <submissions-filter-panel-options
+    v-model="filter"
+    header-label="submissions.filters.status_header"
+    :options="options"
+    :allowed-values="allowedValues"
+  />
 </template>
 
 <script lang="ts">
-interface StatusOption {
-  label: string
-  value: string
-  default?: boolean
-}
+import type { FilterOption } from "./SubmissionsFilterPanelOptions.vue"
 
-const options: StatusOption[] = [
+const options: FilterOption[] = [
   { label: "submission.status.DRAFT", value: "DRAFT", default: false },
   {
     label: "submission.status.INITIALLY_SUBMITTED",
@@ -78,41 +46,12 @@ export const defaultOptions = options.map((s) => s.value)
 </script>
 
 <script setup lang="ts">
-import { computed } from "vue"
-import { useI18n } from "vue-i18n"
+import SubmissionsFilterPanelOptions from "./SubmissionsFilterPanelOptions.vue"
 
-const { t } = useI18n()
 const filter = defineModel<string[]>({ default: () => [] })
 
 interface Props {
   allowedValues?: string[]
 }
-const props = defineProps<Props>()
-
-const tOptions = computed(() =>
-  options
-    .filter(
-      (s) => !props.allowedValues || props.allowedValues.includes(s.value)
-    )
-    .map((s) => ({
-      label: s.label ? t(s.label) : s.label,
-      value: s.value
-    }))
-)
-
-const allowedDefaults = computed(() =>
-  props.allowedValues
-    ? defaultOptions.filter((v) => props.allowedValues!.includes(v))
-    : defaultOptions
-)
-
-function selectAll() {
-  filter.value = [...allowedDefaults.value]
-}
-function selectNone() {
-  filter.value = []
-}
-function invert() {
-  filter.value = allowedDefaults.value.filter((v) => !filter.value.includes(v))
-}
+defineProps<Props>()
 </script>
