+++ client/src/routes/admin/beta-users.vue
@@ -90,9 +90,8 @@ import { useRouter } from "vue-router"
 import QueryTable, {
   type QueryTableColumn
 } from "src/components/tables/QueryTable.vue"
-import NameAvatarCell, {
-  type NameAvatarColumn
-} from "src/components/tables/common/NameAvatarCell.vue"
+import type { NameAvatarColumn } from "src/components/tables/common/NameAvatarCell.vue"
+import { userBaseColumns } from "src/components/tables/common/userColumns"
 import FindUserSelect, {
   type FindUserSelectValue
 } from "src/components/forms/FindUserSelect.vue"
@@ -113,27 +112,7 @@ definePage({
 type BetaUserRow = GetBetaUsersQuery["users"]["data"][number]
 
 const columns: (QueryTableColumn | NameAvatarColumn)[] = [
-  {
-    name: "name",
-    required: true,
-    align: "left",
-    field: (row) => row,
-    component: NameAvatarCell,
-    hideUsername: true,
-    sortable: true
-  },
-  {
-    name: "username",
-    align: "left",
-    field: "username",
-    sortable: true
-  },
-  {
-    name: "email",
-    align: "left",
-    field: "email",
-    sortable: true
-  },
+  ...userBaseColumns,
   {
     name: "actions",
     align: "right",
