+++ client/src/routes/admin/users.vue
@@ -62,9 +62,8 @@ graphql(`
 import QueryTable, {
   type QueryTableColumn
 } from "src/components/tables/QueryTable.vue"
-import NameAvatarCell, {
-  type NameAvatarColumn
-} from "src/components/tables/common/NameAvatarCell.vue"
+import type { NameAvatarColumn } from "src/components/tables/common/NameAvatarCell.vue"
+import { userBaseColumns } from "src/components/tables/common/userColumns"
 import DateTimeCell from "src/components/tables/common/DateTimeCell.vue"
 import {
   GetUsersDocument,
@@ -82,27 +81,7 @@ definePage({
 type UserRow = GetUsersQuery["users"]["data"][number]
 
 const columns: (QueryTableColumn | NameAvatarColumn)[] = [
-  {
-    name: "name",
-    required: true,
-    align: "left",
-    field: (row) => row,
-    component: NameAvatarCell,
-    hideUsername: true,
-    sortable: true
-  },
-  {
-    name: "username",
-    align: "left",
-    field: "username",
-    sortable: true
-  },
-  {
-    name: "email",
-    align: "left",
-    field: "email",
-    sortable: true
-  },
+  ...userBaseColumns,
   {
     name: "created_at",
     align: "left",
